Andrew/UI-UX Issue #8 (#22)
Can now navigate between auth pages and book pages

MainActivity shows WelcomeFragment as default

WelcomeFragment has onClickListeners that transition to SignupFragment or LoginFragment
SignupFragment transition after clicking signup not implemented
LoginFragment has onClickListeners that transition to RecoverPasswordFragment or BookActivity
BookActivity shows OwnedFragment as default

Clicking each button on the bottom of the view will navigate you to a different fragment
BorrowedFragment, MapFragment, OwnedFragment, SearchFragment, SettingsFragment
Each of the fragments listed above has their own child Fragment which changes depending on which sub-tab you click on.
For now, each of the subtabs is using the same fragment GenericListFragment
Fragments are created when you click on the corresponding tab and are hidden when you click on a different tab and shown when you click on the tab again. This means we are not recreating fragments each time we select a tab.

Back button should remember which sub/main tabs you visited and should navigate back to the last tab

Only owned tab has subtabs. When we change from buttons to tabs, some fragment instantiation functionality will be different but I will create a separate issue for that so other issues waiting on this can proceed. If you need subtabs (buttons) in the other pages, you can just copy the code in OwnedFragment and setup buttons in the XML. I am hoping to have the buttons replaced with tabs soon for all the pages.

BookActivity and other Fragments might need to be renamed :( I don't know what to call that section of the app.